Extract more to panels and diagnostics, make sure diagnostics is imported
