+++ __init__.py
@@ -1,2 +1,4 @@
 from .core.main import *
+from .core.panels import *
 from .features.completion import *
+from .features.diagnostics import *
