+++ core/panels.py
@@ -1,4 +1,6 @@
 import sublime
+import sublime_plugin
+
 
 OUTPUT_PANEL_SETTINGS = {
     "auto_indent": False,
@@ -22,3 +24,26 @@ def create_output_panel(window: sublime.Window, name: str) -> sublime.View:
     for key, value in OUTPUT_PANEL_SETTINGS.items():
         settings.set(key, value)
     return panel
+
+
+class LspClearPanelCommand(sublime_plugin.TextCommand):
+    """
+    A clear_panel command to clear the error panel.
+    """
+
+    def run(self, edit):
+        self.view.erase(edit, sublime.Region(0, self.view.size()))
+
+
+class LspUpdatePanelCommand(sublime_plugin.TextCommand):
+    """
+    A update_panel command to update the error panel with new text.
+    """
+
+    def run(self, edit, characters):
+        self.view.replace(edit, sublime.Region(0, self.view.size()), characters)
+
+        # Move cursor to the end
+        selection = self.view.sel()
+        selection.clear()
+        selection.add(sublime.Region(self.view.size(), self.view.size()))
